+++ src/main/java/mate/academy/springboot/datajpa/repository/CategoryRepository.java
@@ -3,6 +3,6 @@ package mate.academy.springboot.datajpa.repository;
 import mate.academy.springboot.datajpa.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    Category getById(Integer id);
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    Category getById(Long id);
 }
